requirements.txt
git+https://github.com/nextcord/nextcord.git
nextcord-pretty-help
pyfiglet
humanfriendly
asyncdagpi
cowpy